Inicio do desenvolvimento do CRUD de serviços

+++ Modelos/Servico.cs
@@ -16,6 +16,6 @@ namespace Models
 
         public bool Ativo { get; set; }
 
-        public float Valor { get; set; }
+        public double Valor { get; set; }
     }
 }
